Consolidate bridges into one file
