--- live.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-  <head>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <title>MetaMask &lt; = &gt; Ledger Bridge</title>
-  </head>
-  <body>
-    <script src="./bundle-live.js" type="text/javascript"></script>
-  </body>
-</html>
