--- main-live.js
@@ -1,8 +1,0 @@
-'use strict'
-import 'babel-polyfill';
-import LedgerLiveBridge from './ledger-live-bridge'
-
-(async () => {
-    const bridge = new LedgerLiveBridge()
-})()
-console.log('MetaMask < = > Ledger Live Bridge initialized!')
