+++ package.json
@@ -27,7 +27,6 @@
     "babel-runtime": "^6.26.0",
     "babelify": "^8.0.0",
     "browserify": "^16.2.2",
-    "buffer": "^5.2.0",
-    "merge-stream": "^2.0.0"
+    "buffer": "^5.2.0"
   }
 }
